Updates and Additions

+++ Exam4s/A01_Selection_Sort.c
@@ -13,7 +13,6 @@ void selectsort(int a[], int n)
 {
     int i, j, min;
     
-
     for(i = 0; i<n-1; i++)
     {
         min = i;
@@ -27,7 +26,7 @@ void selectsort(int a[], int n)
 
 int main()
 {
-    double time;
+    long double time;
     clock_t start, stop;
     int arr[] = {1,6,2,8,3,9,4,1,6,2,8,3,9,4,1,6,2,8,3,9,4};
     int n = sizeof(arr)/sizeof(int);
@@ -36,7 +35,7 @@ int main()
     stop = clock();
     for(int i = 0; i<n; i++)
         printf("%d ",arr[i]);
-    time = ((double)stop-(double)start)/CLOCKS_PER_SEC;
-    printf("\nTime Taken: %f",time);
+    time = ((long double)stop-(long double)start)/CLOCKS_PER_SEC;
+    printf("\nTime Taken: %lf",time);
     return 0;
 }
